(sol2.cfront): CalendarQueue::LocalInit(): SunCC warned that a long was being assigned to an int, so we cast.

+++ trunk/ptclassic/src/domains/de/kernel/CalendarQueue.cc
@@ -121,7 +121,7 @@ double lastTime)
 
     cq_lastTime = lastTime;
     n = (int) (lastTime/cq_interval);
-    cq_lastBucket = n % cq_bucketNum;
+    cq_lastBucket = (int) (n % cq_bucketNum);
     cq_bucketTop = (n+1.5)*cq_interval;
 
 // set up queue size change thresholds
